sql changes [2]
Added a where clause for selfhelp page - will look for cases by user id based off the user in the session. 
Fixes to the mysql database so that user IDs can be linked to cases
Changed the "no data" card when there is no problems as they were displaying upside down previosuly.